Assignment_03_meeting_summarizer/requirements.txt
langchain_core==0.3.63
langchain_openai==0.3.18
pydantic==2.11.5
python-dotenv==1.1.0
